% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clippy_says}
\alias{clippy_says}
\title{A stupid function to make Clippy say hello to you in the terminal}
\usage{
clippy_says(what)
}
\arguments{
\item{what}{String. What should Clippy say?}
}
\description{
A stupid function to make Clippy say hello to you in the terminal
}
